pyproject.toml
[project]
name = "ai-project-planner"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "crewai>=1.3.0",
    "crewai-tools>=1.3.0",
    "openpyxl>=3.1.5",
    "pandas>=2.3.3",
    "plotly>=6.3.1",
    "python-dotenv>=1.2.1",
    "pyyaml>=6.0.3",
    "streamlit>=1.51.0",
]

[dependency-groups]
dev = [
    "pytest>=8.4.2",
    "pytest-cov>=7.0.0",
    "pytest-html>=4.1.1",
]
